+++ CrowdControl/src/main/java/com/elmakers/mine/bukkit/plugins/crowd/Controller.java
@@ -4,11 +4,26 @@ import java.util.List;
 import java.util.Random;
 import java.util.logging.Logger;
 
+import net.minecraft.server.EntityChicken;
+import net.minecraft.server.EntityCow;
+import net.minecraft.server.EntityCreeper;
+import net.minecraft.server.EntityGhast;
+import net.minecraft.server.EntityLiving;
+import net.minecraft.server.EntityPig;
+import net.minecraft.server.EntityPigZombie;
+import net.minecraft.server.EntitySheep;
+import net.minecraft.server.EntitySkeleton;
+import net.minecraft.server.EntitySlime;
+import net.minecraft.server.EntitySpider;
+import net.minecraft.server.EntitySquid;
+import net.minecraft.server.EntityZombie;
+import net.minecraft.server.WorldServer;
+
 import org.bukkit.Location;
 import org.bukkit.Server;
 import org.bukkit.World;
+import org.bukkit.craftbukkit.CraftWorld;
 import org.bukkit.entity.Cow;
-import org.bukkit.entity.CreatureType;
 import org.bukkit.entity.Creeper;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Ghast;
@@ -25,6 +40,7 @@ import org.bukkit.entity.Squid;
 import org.bukkit.entity.Zombie;
 import org.bukkit.event.entity.CreatureSpawnEvent;
 
+import com.elmakers.mine.bukkit.borrowed.CreatureType;
 import com.elmakers.mine.bukkit.plugins.crowd.dao.ControlRule;
 import com.elmakers.mine.bukkit.plugins.crowd.dao.ControlledWorld;
 import com.elmakers.mine.craftbukkit.persistence.Persistence;
@@ -53,17 +69,44 @@ public class Controller
 			case ZOMBIE: return (entity instanceof Zombie);
 			case GHAST: return (entity instanceof Ghast);
 			case SLIME: return (entity instanceof Slime);
-			case GIANT_ZOMBIE: return (entity instanceof Giant);
+			case GIANT: return (entity instanceof Giant);
 		}
 
 		return false;
 	}
 	
-	protected LivingEntity spawn(Location location, CreatureType type)
+	protected EntityLiving spawn(Location location, CreatureType type)
 	{
-		return server.spawn(location, type);
+		EntityLiving e = null;
+		CraftWorld craftWorld = (CraftWorld)location.getWorld();
+		WorldServer world = craftWorld.getHandle();
+
+		switch (type)
+		{
+			case SHEEP: e = new EntitySheep(world); break;
+			case PIG: e = new EntityPig(world); break;
+			case CHICKEN: e = new EntityChicken(world); break;
+			case COW: e = new EntityCow(world); break;
+			case CREEPER: e = new EntityCreeper(world); break;
+			case PIG_ZOMBIE: e = new EntityPigZombie(world); break;
+			case SKELETON: e = new EntitySkeleton(world); break;
+			case SPIDER: e = new EntitySpider(world); break;
+			case SQUID: e = new EntitySquid(world); break;
+			case GHAST: e = new EntityGhast(world); break;
+			case ZOMBIE: e = new EntityZombie(world); break;
+			//case GIANT_ZOMBIE: e = new EntityGiantZombie(world); break;
+			case SLIME: e = new EntitySlime(world); break;
+			//case FISH: e = new EntityFish(world); break;
+		}
+
+		if (e != null)
+		{
+			e.c(location.getX(), location.getY(), location.getZ(), location.getYaw(), 0.0F);
+	        world.a(e);
+		}
+		return e;
 	}
-	
+
 	public void controlSpawnEvent(ControlledWorld world, CreatureSpawnEvent event)
 	{
 		List<ControlRule> rules = world.getRules();
@@ -103,9 +146,10 @@ public class Controller
 		}
 	}
 	
-	public int nuke(World world, CreatureType entityType, boolean nukeAll)
+	public int nuke(ControlledWorld targetWorld, CreatureType entityType, boolean nukeAll)
 	{
 		int killCount = 0;
+		World world = targetWorld.getId().getWorld(server);
 		List<LivingEntity> entities = world.getLivingEntities();
 		for (LivingEntity entity : entities)
 		{
@@ -121,6 +165,6 @@ public class Controller
 	
 	private Server server;
 	private final Random rand = new Random();
-	private static boolean debugLog = false;
+	private static boolean debugLog = true;
 	private static final Logger log = Persistence.getLogger();
 }